+++ cmd/neofs-node/object.go
@@ -11,7 +11,6 @@ import (
 	"sync/atomic"
 	"time"
 
-	"github.com/google/uuid"
 	lru "github.com/hashicorp/golang-lru/v2"
 	iec "github.com/nspcc-dev/neofs-node/internal/ec"
 	coreclient "github.com/nspcc-dev/neofs-node/pkg/core/client"
@@ -627,16 +626,16 @@ func (x storageForObjectService) VerifyAndStoreObjectLocally(obj object.Object) 
 	return x.putSvc.ValidateAndStoreObjectLocally(obj)
 }
 
-func (x storageForObjectService) GetSessionPrivateKey(usr user.ID, uid uuid.UUID) (ecdsa.PrivateKey, error) {
-	k, err := x.keys.GetKey(&util.SessionInfo{ID: uid, Owner: usr})
+func (x storageForObjectService) GetSessionPrivateKey(account user.ID) (ecdsa.PrivateKey, error) {
+	k, err := x.keys.GetKey(&account)
 	if err != nil {
 		return ecdsa.PrivateKey{}, err
 	}
 	return *k, nil
 }
 
-func (x storageForObjectService) GetSessionV2PrivateKey(issuer user.ID, subjects []sessionv2.Target) (ecdsa.PrivateKey, error) {
-	k, err := x.keys.GetKeyBySubjects(issuer, subjects)
+func (x storageForObjectService) GetSessionV2PrivateKey(subjects []sessionv2.Target) (ecdsa.PrivateKey, error) {
+	k, err := x.keys.GetKeyBySubjects(subjects)
 	if err != nil {
 		return ecdsa.PrivateKey{}, err
 	}